requirements.txt
numpy==1.22.0
pandas==1.3.5
requests==2.27.1
openpyxl==3.1.0
spacy~=3.5.1
textstat~=0.7.3
Markdown~=3.4.1
beautifulsoup4~=4.12.0
toml~=0.10.2
textblob~=0.17.1
pycountry~=22.3.5
loguru~=0.7.0
python-dotenv~=1.0.0